requirements.txt
libtmux==0.37.0
GitPython==3.1.43
pydantic==2.8.2
rich==13.7.1
pytest==8.3.2